TVG-102 Blueprints for Reminders endpoints
