+++ api_blueprints/__init__.py
@@ -1,2 +1,3 @@
+from api_blueprints.reminders import reminder_blueprint, reminders_blueprint
 from api_blueprints.search_items import search_items_blueprint
 from api_blueprints.shows import shows_blueprint
